requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.25.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
pandas>=2.1.0
plotly>=5.18.0
tenacity>=8.0.0
